backend/requirements.txt
numpy>=1.24
requests>=2.31
tqdm>=4.65
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4